Update code using toIPv6/fromIPv6 functions

+++ Network/DNS/Encode.hs
@@ -121,7 +121,7 @@ encodeRR ResourceRecord{..} =
 encodeRDATA :: RDATA -> SPut
 encodeRDATA rd = case rd of
     (RD_A ip)          -> mconcat $ map putInt8 (fromIPv4 ip)
-    (RD_AAAA ip)       -> mconcat $ map putInt16 (fromIPv6 ip)
+    (RD_AAAA ip)       -> mconcat $ map putInt8 (fromIPv6 ip)
     (RD_NS dom)        -> encodeDomain dom
     (RD_CNAME dom)     -> encodeDomain dom
     (RD_DNAME dom)     -> encodeDomain dom
